Implement support for segment package 'name' metadata. MILESTONE.
You can now add an arbitrary 'name' to the metadata of any segment package.
Doing this controls how the segment appears everywhere in the system.

MILESTONE: the primary benefit here is that you can now name segments things
like A, B, C or A1, A2, A3 and so on. These are probably the most natural names
for segments.

This commit renames the Red Example Score segments A, B, C.

Example:

    Red Example Score (2013) - segments

       1: A
       2: B
       3: C

          segments - copy (cp)
          segments - new (new)
          segments - remove (rm)
          segments - rename (ren)

    >