Implementing special (**) menu. MODEL.
MODEL: The (**) command will take the composer to a special view of the system
that is truly global: all scores and all asset libraries will be visible from
(**). (The provisional name for this view is simply 'the' library.)

MODEL: In the library composers will be able to run starred commands like
(no**) to open literally ALL __init__.py files in the system or (mdls**) to
list literally ALL of the __metadata__.py files in the system.

MODEL: Note that starred commands are of course already available in the menus
provided by MaterialPackageWrangler, SegmentPackageWrangler,
ScorePackageWrangler. Calling (no**) against the SegmentPackageWrangler opens
the __init__.py files in all segments; calling (no**) in "the" library will
open all __init__.py files EVERYWHERE.

The take-away here is that (**) will afford a maximally global type of command
iteration.