Added Session.is_pending_redraw. RENDERING.
The following series of commits are design to simply scren redraws.

This commit adds a single session variable than can be set anywhere in the
system anytime the composer enters a command that should result in a redraw.
The Menu class will later notice this setting and effect the redraw.