Implemented PitchRange.from_pitches().
PitchRange.from_pitches() is a class method that functions as an alternative
initializer in the style of dict.fromkeys().

Note that we spell 'from_keys' instead of 'fromkeys' because custom classes
need not use "Python case" for method names.

This is the first commit in a two-part commit that will simplify the PitchRange
initializer and (finally) make it keyword-only and completely explicit.