Added stub 'maker.py - autoedit' (mae) command. MODEL.
MODEL: a new realization indicates that there are probably two types of maker:

    * MAKER TYPE I: __call__() takes arguments
    * MAKER TYPE II: __call__ takes no arguments

Type I is what we're used to dealing with: rhythm-makers all get called on an
arbitrary list of divisions.

Type II may not yet exist in the mainline but exists in a number of my own
scores. ZaggedPitchClassMaker constitutes an example.

Here's how you instantiate a zagged pitch-class maker:

    maker = ZaggedPitchClassMaker(
        pc_cells=(
            [8, 4, 3, 2, 11],
            [5, 4, 6, 8, 7],
            [9, 6, 5, 0, 11, 10],
            ),
	division_cells=(
            [[1], [1], [1], [1, 1]],
            [[1], [1], [1], [1, 1, 1], [1, 1, 1]],
            ),
	grouping_counts=(1, 1, 2, 3),
        )

Here's how you call a zagged pich-class maker:

    maker()

The maker outputs a cyclic payload tree of pitch-classes.

So consider the following:

1. ZaggedPitchClassMaker is an example of a maker whose __call__() method takes
no input arguments. (A type II maker, according to the distinction above).

2. The effort of working with a ZaggedPitchClassMaker is all initialization
work: the composer has to work through potentially many combinations of pc
cells, division cells and grouping counts to get precisely the right type of
nested pitch-class output.

3. Autoediting makes sense here. It makes sense to autoedit a
ZaggedPitchClassMaker just like it makes sense to autoedit a TaleaRhythmMaker.
(In fact it probably makes sense to autoedit all makers because one of the
features in common to all makers is a significant initializer signature).

4. Autoediting the *output* of a ZaggedPitchClassMaker doesn't make any sense.
(Rather it's autoediting the input of a ZaggedPitchClassMaker that makes
sense).

So what should be the artifacts in a material package that features a
ZaggedPitchClassMaker?

It seems like we should allow for a new maker.py file in the material package
that will house something like ...

    maker = ZaggedPitchClassMaker(
        pc_cells=(
            [8, 4, 3, 2, 11],
            [5, 4, 6, 8, 7],
            [9, 6, 5, 0, 11, 10],
            ),
	division_cells=(
            [[1], [1], [1], [1, 1]],
            [[1], [1], [1], [1, 1, 1], [1, 1, 1]],
            ),
	grouping_counts=(1, 1, 2, 3),
        )

... and will be autoeditable.

Then we'll need a command like 'maker.py - interpret' (mi) to interpret the
maker, call it, and write the output to the output.py file.

This means that the composer could use a maker.py file in lieu of a
definition.py file.

These changes will be made in future commits.

The present commit adds the (mae) commands as a stub.