Fixed Git (rrv).
Explanation: the 'repositor - revert' (rrv) command is modeled on the
Subversion 'revert' command. The purpose of the command is to get rid of any
changes in the local directory (and in its subdirectories). The best way to do
this in Git is with 'git checkout'.

This commit changes the Git implementation of (rrv) to use 'git checkout'
instead of 'git reset'.

Note that 'git reset' is preserved in the IDE. The IDE uses 'git reset' to
clean up after adding test artifacts to the repository.