Implemented 'definition.py - autoedit' (dae). NEW.
Description: the (dae) command allows you to treat a material package
definition.py file as the static container of an autoeditable class.

This means that it's now possible to make a material package and
then use an autoeditor to create, for example, a customized Piano or a fully
populated TempoInventory all in only a few keystrokes.

The storage format of the target is then written cleanly into the definition.py
file.

This is a new and very general way of working with material definition files.