Improved (dae).
You can now use (dae) followed by (dp).

The combination will let you autocreate a complex class (and store it in the
definition.py file) and then publish the storage format of the class the
output.py file.

This is a way of working with material packages that involves writing no code
by hand. Everything is done in the UI.