Added support for autoadvance toggling. Suffix (/) to toggle. NEW.
Description. Consider a class with many properties, like an instrument. When
first defining the instrument autoadvance is very useful because you don't have
to press (1) to edit the first property, followed by (2) to edit the second
property, followed by (3) to edit the third property. But when making surgical
changes to a single instrument property after definition, it makes sense to
just type (3) to edit the third property. So clearly some sort of control
should be given to the composer to toggle autoadvance on and off.

This commit adds autoadvance toggle in the form of slash-suffixed commands.
Just type (1/) to edit the first property and then be autoadvanced to
properties 2, 3, 4 and 5. When you're done editing you can turn autoadvance off
by entering (my last value/) or just by entering (/).

Example: autoadvances properties 2 through 6. Notice the slashes entered by the
composer to initiate input and then terminate input:

    Abjad IDE - materials - test viola (Bača) (EDIT)

       1: clefs (cf): alto, treble
       2: instrument name (in): viola
       3: instrument name markup (im): \markup { Viola }
       4: range (rg): [C3, D6]
       5: short instrument name (sn): va.
       6: short instrument name markup (sm): \markup { Va. }
       7: sounding pitch of written middle C (sp): c'

          done (done)
          set to none (none)

    |> 2/
    Instrument name]> fancy viola
    Instrument name markup]> Markup('fancy viola')
    Pitch range]> [C3, C7]
    Short instrument name]> fn. va.
    Short instrument name markup]> Markup('fn. va.')/

    Abjad IDE - materials - test viola (Bača) (EDIT)

       1: clefs (cf): alto, treble
       2: instrument name (in): fancy viola
       3: instrument name markup (im): \markup { fancy viola }
       4: range (rg): [C3, C7]
       5: short instrument name (sn): fn. va.
       6: short instrument name markup (sm): \markup { fn. va. }
       7: sounding pitch of written middle C (sp): c'

          done (done)
          set to none (none)

    |>