Added provisional support for composer-initiated (@) autostart. NEW.
Description. You will soon be able to enter (da@) for 'definition.py -
autoedit' with the autoeditor turned on and automatically prompting you for
a value for the first property of the target. Right now you can enter (da) to
start the autoeditor on the target ... but you then need to enter (1) or
equivalent to begin editing the different properties of the target. Using (da@)
will be a short cut that says to start absolutely everything immediately.