Extended TestManager.compare_pdfs().
The method now discards the first binary stream object contained in the PDF.

Why?

Description: consider the output.ly file for Red Example Score segment A. Then
consider what happens if you call (oli) repeatedly on this file to read
output.ly and produce output.pdf. The behavior is surprising: three out of four
times output.pdf will compare equal to itself; but the fourth version of the
PDF will compare unequal to previous versions of itself. Using "diff -a" shows
that the difference always lies in the first binary stream object contained in
the PDF. A cursory look at the PDF spec hosted on the Adobe's website, doesn't
yield any easy answer as to why this should be. So this commit takes the naive
approach of simply discarding the first binary stream object contained in every
PDF passed in for comparison. The advantage is that the approach seems to work.
The disadvantage may turn out to be that visually significant information could
be housed in the first binary stream object in a PDF. But the changes committed
here are pushed because this doesn't appear to be the case in practice.