Harmonized commands.
Use (??) to inspect commands in your favorite menus.

The basic idea of this commit is that two- and three-character verbs (like ls
and rm) can be abbreviated to a single character in compound commands.
So '__init__.py - list' (nls) changed to (nl) and '__metadata__.py - list'
changed to (mdl). And so on.