Updated two Subversion-related tests.
Subversion-related tests all proceed the same way: first find an up-to-date
Subversion-managed directory on the composer's system. If the composer
maintains no Subversion repositories then no up-to-date Subversion-managed
directory will be found. This is ok and the protocol is to then exit the test
without proceeding any further.

The commit here updates two tests to follow that protocol.