Implemented a basic Selection.__illustrate__().
The method works by attempting to evaluate the storage format of the selection.
(This will have the effect of copying the contents of the selection in a way
that severs all references from the source score from which the selection was
made.)

If it works to evaluate the storage format of the selection, then the method
tries to wrap the resulting music in a staff. If this works then the method
checks to see if all pitches equal C4. If this is the case then the context
name of the staff is set to 'RythmicStaff'.

The intent is to allow simple snippets of single-voice music to work with
show(). And this does indeed work.

Example:

    >>> staff = Staff("c'4 d'4 e'4 f'4 g'4 a'4 b'4 c''4")

    >>> staff[2:6]
    SliceSelection(Note("e'4"), Note("f'4"), Note("g'4"), Note("a'4"))

    >>> selection = _

    >>> lilypond_file = selection.__illustrate__()

    >>> f(lilypond_file)
    % 2014-07-08 16:21

    \version "2.19.9"
    \language "english"

    \header {
    	tagline = \markup {}
    }

    \layout {}

    \paper {}

    \score {
    	\new Score <<
    		\new Staff {
    			e'4
    			f'4
    			g'4
    			a'4
    		}
    	>>
    }