Introduced "depot" for (U), (D), (K), (M), (G), (Y). TERMINOLOGY.
IDE now distinguishes systematically between (m) for 'materials directory' and
(M) for 'materials depot'.

This provides a name for the menuing provided by (M) and friends. The word
'depot' differs enough from 'view' to keep the two concepts clear: it makes
sense to apply a view to the materials depot. And it makes just as much sense
to apply a view to the materials directory.