Changed experimental/musicmakertools to experimental/newmusicmakertools.
The package is currently empty. But the package will soon house new classes.

The name change committed here anticipates the eventual addition of a
musicmakertools to the main Abjad API. (Which will mean that musicmakertools
and rhythmmakertools would model similarly.)

The name change committed here harmonizes the names of these three package ...

    * experimental.newabjadbooktools
    * experimental.newindicatortools
    * experimentla.newmusicmakertools

... too.