Added experimental newmusicmakertools.MeasurewiseDivisionMaker class.
PURPOSE: MeasurewiseDivisionMaker will function to change measures into
divisions.

BACKGROUND: Metric music divides measures into beats according to more-or-less
universally understood rules. (See the logic implemented in
MutationAgent.rewrite_meter() for the Abjad implementation of these rules.)
But nonmetric music divides measures into divisions in arbitrarily many ways.
A related family of division-makers will model the different ways composers
might want to divide measures into divisions.

USAGE: MeasurewiseDivisionMaker will follow the two-step configure-once /
call-repeatedly pattern all rhythm-makers implement already. Composers will
call a configured measurewise division-maker with an arbitrary list of measures
(or time signatures) and the maker will return a list of divisions.

CONCORDANCE: the concept of the division-maker will supersede the logic
implemented in TimeContiguousSetMethodMixin.set_divisions() in
musicexpressiontools. The purpose of statements like ...

    red_segment.set_divisions([(1, 8), (3, 8)])

... was to allow composers to specify the way in which divisions would be
created for all or part of a segment. Moving forward this will be done by
configuring a division-maker, passing the configured division-maker to a
(larger-scale) music-maker and then adding the configured music-maker into an
(even larger) segment-maker.

MODEL: it is possible that the expanding family of makers (rhythm-makers,
segment-makers and now division-makers and soon other music-makers) will be the
eventual fate of what became of most the thinking in musicexpressiontools.

STATUS: The concept of a division-maker is new as of this commit.
MeasurewiseDivisionMaker is currently a stub. MeasurewiseDivisionMaker is the
first class committed to the newmusicmakertools package.

TERMINOLOGY: the use of 'measurewise' in MeasurewiseDivisionMaker is taken
from these classes ...

    quantizationtools.BeatwiseQSchema
    quantizationtools.BeatwiseQSchemaItem
    quantizationtools.BeatwiseQTarget
    quantizationtools.MeasurewiseAttackPointOptimizer
    quantizationtools.MeasurewiseQSchema
    quantizationtools.MeasurewiseQSchemaItem
    quantizationtools.MeasurewiseQTarget

... in quantizationtools.