Updated experimental.Division tests.
Preparing Division for migration to mainline durationtools.

This commit adds the following asserts to the Division constructor ...

    assert self.is_left_closed
    assert not self.is_left_open

... which are designed to enforce the idea that all divisions will be closed /
open.

MODEL: *ALL* timespans established by the gylphs of received notation in score
are closed / open. We've known this for a while. The decision here is to
enforce this fact on Division prior to migration to the mainline. This means
that Division will no longer inherit from BoundedObject.

A following commit will enforce the second half of this: all divisions will be
right-open.