Extended TieSpecifier.tie_across_divisions. NEW FUNCTIONALITY.
TieSpecifier implements two properties:

    * TieSpecifier.tie_across_divisions
    * TieSpecifier.tie_split_notes

This commit concerns TieSpecifier.tie_across_divisions.

OLD: TieSpecifier.tie_across_divisions could be set to either true or false.
When true ties would be applied across the boundaries of all divisions. When
false no ties would be applied across the boundaries of divisions. (The
interpretation of the property was all or nothing, as the boolean implied.)

NEW: TieSpecifier.tie_across_divisions can now be set to a tuple or list of
boolean-equivalent values like [0, 1, 1] or [1, 0] or [0, 0, 1, 1, 1, 0, 1].
The iterable of booleans is interpreted as a cyclic pattern showing which pairs
of divisons should be tied (true-valued) or not tied (false-valued).

See the EvenRunRhythmMaker API entry under 'tie_specifier' for a complete set
of examples.