Added TupletSpellingSpecifier to all rhythm-makers.
OLD: only TupletRhythmMaker implemented a tuplet_spelling_specifier property.

NEW: all rhythm-makers now inherit a tuplet_spelling_specifier property from
the base RhythmMaker class.

MODEL: the realization is that basically all rhythm-makers have the possibility
of creating tuplets and so all rhythm-makers should afford a spelling interface
for the composer. (Consider that NoteRhythmMaker happily generates a tuplet
monda when passed a duration like 1/12 for example.)

STATUS: the tuplet_spelling_specifier property is not yet integrated into the
family of rhythm-makers. This commit serves only to add the property to the
family of rhythm-makers.