Added 'preferred_denominator_from_divisions' property to TupletRhythmMaker.
NEW FUNCTIONALITY: set the property to true to force tuplet ratio (where
possible) to be expressed in terms of the numerators of input durations.

The old default behavior is that tuplet ratios are always reduced and that
the numerator and denominator of each tuplet are always relatively prime (thus
making only ratios like 5:4 and 3:2 possible).

With the new behavior turned on the numerator and denominator of each tuplet
may not be relatively prime (thus making ratios like 10:8 and 6:4 possible).

See the TupletRhythmMaker API entry for examples.