Changed BowSpanner._get_bowing_indicators() to _get_annotations().
The new name seems less specific but will harmonize with a method of the same
name on TempoSpanner (and conceivably on all spanners that format in this new
way).

MODEL: we will eventually need a word for "indicators that do not contribute
formatting but that are instead analyzed by a smart spanner".

The use of 'annotation' in this commit kinda serves to do this. Even though we
already have an Annotation class in the system.