Impelemented DurationSpellingSpecifier 'spell_metrically' property.
The property is currently a stub.

When set to true the specifier tells rhythm-makers to break each input division
down according to the common practice understanding of meter. (So 3/4 and 6/8
will be broken down in two different ways.)

When set to false the specifier tells rhythm-makers to continue spelling
durations with the fewest number of glyphs. (Old behavior.)

Defaults to false.