Added Selector TODO items. MODEL.
Selector's inequality-related methods can be replaced in favor ...

    Selector.by_duration(duration=None, ...)
    Selector.by_length(length=None, ...)

... where the 'duration' and 'length' parameters are generalized to accept new
DurationInequality and LengthInequality objects.

The DurationInequality and LengthInequality classes don't exist yet.

The DurationInequality and LengthInequality classes can be modeled on ...

    timespantools.SimpleInequality
    timespantools.CompoundInequality

... though those two classes should not be used as the literal base classes for
the new DurationInequality and LengthInequality classes (because there's extra
stuff in the timespan classes that's not appropriate for the two new classes).

MODEL: this is basically a pattern of object-modeled inequalities. This is
powerful for us. Because there are a number of functions and methods throughout
the system the interfaces to which have always been less than perfect because
the set of possible inequalities that might be used as restrictions in some
operation is not bounded. (Note with Selector that it's more or less arbitrary
that there are no Selector.greater_than_or_equal(),
Selector.Selector.more_than_or_equal() and so on.) The solution is allow thin,
core classes that allow the modeling of any inequality as an object that can be
passed as input to other composite classes.