Extended BurnishSpecifier. INPUT.
OLD:

    >>> burnish_specifier = rhythmmakertools.BurnishSpecifier(
    ...     lefts=[Rest, Note],
    ...     left_lengths=[1],
    ...     )

NEW:

    >>> burnish_specifier = rhythmmakertools.BurnishSpecifier(
    ...     lefts=[Rest, Note],
    ...     left_lengths=[1],
    ...     )

You can now pass in classes (Note, Rest) as the arguments to
BurnishSpecifier.lefts and BurnishSpecifier.rights.

This makes the function of these two properties considerably more
self-documenting.

CHANGE: should be noted in the 2.16 release notes.

Note that the old integer notation (with -1 for Rest and 1 for Note) still
works. Though passing in classes is now the preferred way of configuring the
class for reasons of documentation.