Added TaleaRhythmMaker.output_mask property. NEW FUNCTIONALITY.
Set like ...

    output_mask=[0, 0, 1]

... to a boolean vector.

The rhythm-maker will then read the output mask cyclically as a final step in
rhythmic creation. Each 0 encountered forces the output division to a rest.
Each 1 encountered preserves the output division as-is.

Useful for punching holes in rhythmic output in a patterned way.

Note that it makes sense to implement this functionality on all rhythm-makers.

This commit implements against only TaleaRhythmMaker.

See the API entry for examples.