Improved DurationSelectorCallback.
DurationSelectorCallback.inequality now accepts a tuple of one or more of
None, Exact, More or Less. This allows us to model model the concept of "select
all components whose duration is greater or equal to x".

The change is provisional until a proper DurationInequality class is
implemented, after the current round of score are complete.