Implemented Glissando.include_tied_leaves property.
OLD: the Glissando spanner always included leaves in the middle of a logical
tie. This is almost certainly not the behavior normally desired.

NEW: the Glissando spanner now glisses into the first leaf in a logical tie and
out of the last leaf of a logical tie (omitting leaves internal to a logical
tie), by default. The old behavior is available by initializing a glissando
with include_tied_leaves=True.

See the Glissando API entry for examples.