Added developer property to Instrument.
Added private _do_not_format property.

No public interface is currently implemented.

But developers can set the property to true to cause any instrument
to suppress LilyPond output.

This can be useful when you want an instrument attached to score but you do not
want the LilyPond instrumentName and shortInstrumentName settings to be written
to output.

Kind of the opposite of annotation.