Added scoremanager/test/autoeditor directory.
The directory houses files like this:

    test_Clef_autoedit.py
    test_Instrument_autoedit.py
    test_Markup_autoedit.py
    test_NoteAndChordsHairpinHandler_autoedit.py
    test_NumberedPitch_autoedit.py
    test_PatternedArticulationsHandler_autoedit.py
    test_Performer_autoedit.py
    test_PitchRange_autoedit.py
    test_RegistrationComponent_autoedit.py
    test_ReiteratedArticulationHandler_autoedit.py
    test_ReiteratedDynamicHandler_autoedit.py
    test_TaleaRhythmMaker_autoedit.py
    test_Tempo_autoedit.py
    test_TerracedDynamicsHandler_autoedit.py
    test_noteAndChordHairpinHandler_autoedit.py

The files are named one per class.

The purpose of each file is to ensure the autoeditability of a given class.

It is conceivable that the directory could eventually house one file for each
mainline class (as we work to make each mainline class autoeditable). This
potentially large number of files is the reason for the separate directory
committed here.

It is also possible that in a later release of Abjad that these test files will
migrate into the mainline and be housed with the classes they test, just like
normal test files. But the reason for keeping these test files in the
scoremanager branch is because they depend on the IDE. Mainline tests should
not depend on the IDE. (At least not until the IDE becomes part of the
mainline.)